修复文字图片带有透明度，预乘出现的黑色描边问题。

+++ yyevac/src/main/cpp/bean/effect.h
@@ -6,8 +6,8 @@
 
 class Effect {
 public:
-    int effectWidth;
-    int effectHeight;
+    float effectWidth;
+    float effectHeight;
     int effectId;
     std::string effectTag;
     std::string effectType;
