+++ yyevac/src/main/cpp/util/vertexutil.cpp
@@ -7,28 +7,28 @@
 #define LOG_TAG "VertexUtil"
 #define ELOGE(...) yyeva::ELog::get()->e(LOG_TAG, __VA_ARGS__)
 
-float* VertexUtil::create(int width, int height, std::shared_ptr<PointRect> rect, float *array) {
+float* VertexUtil::create(float width, float height, std::shared_ptr<PointRect> rect, float *array) {
     if (array == nullptr) {
         ELOGE("array is null");
         return array;
     }
     //x0
-    array[0] = switchX((float)rect->x / width);
+    array[0] = switchX(rect->x / width);
     //y0
-    array[1] = switchY((float)rect->y / height);
+    array[1] = switchY(rect->y / height);
     //x1
-    array[2] = switchX((float)rect->x / width);
+    array[2] = switchX(rect->x / width);
     //y1
-    array[3] = switchY(((float)rect->y + rect->h) / height);
+    array[3] = switchY((rect->y + rect->h) / height);
 
     //x2
-    array[4] = switchX(((float)rect->x + rect->w) / width);
+    array[4] = switchX((rect->x + rect->w) / width);
     //y2
-    array[5] = switchY( (float)rect->y / height);
+    array[5] = switchY( rect->y / height);
     //x3
-    array[6] = switchX(((float)rect->x + rect->w) / width);
+    array[6] = switchX((rect->x + rect->w) / width);
     //y3
-    array[7] = switchY(((float)rect->y + rect->h) / height);
+    array[7] = switchY((rect->y + rect->h) / height);
 
     return array;
 }
